+++ CHANGELOG.md
@@ -13,6 +13,7 @@
     * Fix: File downloading.
     * Fix: Allow to select default variation.
     * Fix: Variation sale datepickers.
+    * Fix: Improve attachments structure.
 * 2.0.7 - 2016.12.28:
     * Fix: Downloadable email link.
     * Fix: Updating shipping methods in checkout.